+++ app/src/main/AndroidManifest.xml
@@ -22,14 +22,14 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="poly.darkdepths.strongbox.Welcome"
+            android:name=".miscGUI.Welcome"
             android:label="@string/app_name" >
         </activity>
         <activity
             android:name="poly.darkdepths.strongbox.CameraActivity"
             android:label="@string/app_name" >
         </activity>
-        <activity android:name=".Unlock"
+        <activity android:name=".miscGUI.Unlock"
             android:label="Strongbox" >
         </activity>
     </application>
