+++ app/src/main/java/poly/darkdepths/strongbox/io/IOCipherFileChannelWrapper.java
@@ -1,4 +1,4 @@
-package poly.darkdepths.strongbox;
+package poly.darkdepths.strongbox.io;
 
 /**
  * Created by poly on 3/28/15.
